fs: f2fs: avoid needless checkpoint during fsync
MIUI-1728660

This patch enhances the performance of fsync.

Doing checkpoint every time when setting xattr for dir can seriously
affect the performance of fsync. By adding the inode to the tracking
list when setting up xattr, we can avoid doing checkpoint when xattr
of the parent directory is not set.

Signed-off-by: liuchao12 <liuchao12@xiaomi.com>
Change-Id: I6e5bc4065dfc804751dc915e7972c2c701ef9912
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>